pyproject.toml
[project]
name = "privacy-report-generator"
version = "0.1.0"
description = "Generate privacy reports from data flow analysis"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "tree-sitter>=0.23.0",
    "tree-sitter-arkts-open>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent", "src/memory", "src/tools", "src/utils"]
